refactor: implement build_croissant protocol on all FileTypeHandlers
Move all format-specific logic out of MetadataGenerator into each handler.
The generator now owns only FileObject creation and ID assignment; handlers
own their RecordSet and FileSet construction via the new build_croissant
abstract method.

- Add build_croissant @abstractmethod to FileTypeHandler base class
- Implement build_croissant on ImageHandler, CSVHandler, ParquetHandler,
  WFDBHandler
- Move _build_fields (Arrow schema -> mlc.Field, handles nested types) to
  handlers/utils.py
- MetadataGenerator: 647 -> 249 lines, zero format-specific key checks
- Rename file_objects -> distributions to match Croissant spec terminology

Adding a new format now requires only a new handler + registry entry.
